have the ntquery at one level working

+++ src/main.rs
@@ -2,7 +2,7 @@ use jwalk::WalkDir;
 use std::env;
 use std::time::Instant;
 
-use directory_walking::ntqeurydirectoryfile::dir_walk;
+use directory_walking::ntquerydirectoryfile::dir_walk;
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -18,7 +18,7 @@ fn main() {
     println!("The ntquerydirectoryfile time {:?}, and the files {:?}", end.duration_since(start), count);
 }
 
-fn run_jwalk(directory: &str) -> u32 {
+fn run_jwalk(directory: &str) -> usize {
     let mut count = 0;
     for _ in WalkDir::new(directory).sort(true) {
         count += 1;
